adds provision for sending data to the API
